extract collectible types to a definition file
