+++ src/store/index.ts
@@ -37,7 +37,7 @@ import incomingTransactions, {
 } from 'src/logic/safe/store/reducer/incomingTransactions'
 import safe, { SAFE_REDUCER_ID } from 'src/logic/safe/store/reducer/safe'
 import transactions, { TRANSACTIONS_REDUCER_ID } from 'src/logic/safe/store/reducer/transactions'
-import { NFTAssets, NFTTokens } from 'src/logic/collectibles/sources/OpenSea'
+import { NFTAssets, NFTTokens } from 'src/logic/collectibles/sources/collectibles'
 import { SafeReducerMap } from 'src/routes/safe/store/reducer/types/safe'
 import allTransactions, { TRANSACTIONS, TransactionsState } from '../logic/safe/store/reducer/allTransactions'
 import { AddressBookState } from 'src/logic/addressBook/model/addressBook'
